[create]: Guest Book CRUD

+++ activity-service/src/main/java/com/example/activityservice/jpa/RunningRepository.java
@@ -3,5 +3,5 @@ package com.example.activityservice.jpa;
 import com.example.activityservice.DAO.RunningDao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RunningRepository extends JpaRepository<RunningDao, Integer> {
+public interface RunningRepository extends JpaRepository<RunningDao, Long> {
 }
