+++ activity-service/src/main/java/com/example/activityservice/jpa/SleepRepository.java
@@ -3,5 +3,5 @@ package com.example.activityservice.jpa;
 import com.example.activityservice.DAO.SleepDao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SleepRepository extends JpaRepository<SleepDao, Integer> {
+public interface SleepRepository extends JpaRepository<SleepDao, Long> {
 }
